SA and FG Costing

+++ CommonModels/WeightedAvgRate.cs
@@ -4,8 +4,8 @@
     {
        public int Id { get; set; }  
         public string? Itemnumber { get; set; }
-        public int totalbalance_qty { get; set; }
-        public int Avg_cost { get; set; }
+        public decimal totalbalance_qty { get; set; }
+        public decimal Avg_cost { get; set; }
         public DateTime? update_date { get; set; }
         public TimeSpan? updated_time { get; set; }
